+++ examples/chat/index.html
@@ -2,7 +2,7 @@
 <html>
 	<head>
 		<meta name="viewport" content="width=device-width, initial-scale=1">
-		<script src="../../gun.js"></script>
+		<script src="../../dist/gun.js"></script>
 	</head>
 	<body>
 		<style>
