+++ examples/todo/index.html
@@ -5,7 +5,7 @@
 		<form id="add"><input id="todo"/><button>Add</button></form>
 		<ul id="todos"></ul>
 
-		<script src="../../gun.js"></script>
+		<script src="../../dist/gun.js"></script>
 		<script> // minimal jQuery polyfill
 			var $ = document.querySelector.bind(document);
 			function clean(text){ return String(text).replace(/\</ig, '&lt;') }
