+++ lib/server.js
@@ -1,8 +1,8 @@
-;(function(){
+// ;(function(){
 	console.log("Hello wonderful person! :) I'm mark@gunDB.io, message me for help or with hatemail. I want to hear from you! <3");
 	var Gun = require('../dist/gun');
 	require('./s3');
 	require('./file');
 	require('./wsp');
 	module.exports = Gun;
-}());
+// }());
