+++ src/is/base.js
@@ -1,7 +1,7 @@
 /**
  * Created by Paul on 9/8/2016.
  */
-import Utils from '../utilities';
+import Utils from '../utilities/base';
 import Obj from '../utilities/obj';
 import Text from '../utilities/text';
 import Reserved from '../reserved';
