+++ src/utilities/index.js
@@ -1,23 +1,11 @@
 /**
- * Created by Paul on 9/7/2016.
+ * Created by Paul on 9/8/2016.
  */
 
-import List from './definitions/list'
+import base from './base';
+import Text from './text'
+import List from './list'
+import Obj from './obj'
+import Time from './time'
 
- let fns = {
-    is: function (fn) {
-      return (fn instanceof Function);
-    }
-  };
- let bi = {
-    is: function (b) {
-      return (b instanceof Boolean || typeof b == 'boolean');
-    }
-  }
-let num = {
-    is: function (n) {
-      return !List.is(n) && (Infinity === n || n - parseFloat(n) + 1 >= 0)
-    }
-  }
-
-export default {fns, bi, num};
+export default {...base, Text, List, Obj, Time };
